[FIX] 작성하기 / 검색기록 저장-조회 로직 분리
